refactor: cli를 command로 수정
