+++ src/command/views/InputView.js
@@ -10,7 +10,6 @@ export default class InputView {
 
     Validator.validatePurchaseAmount(purchaseAmount);
 
-    return purchaseAmount;
   }
 
   static async readWinNumbers() {
